LRQA-42795 Move logic for finding supplementary portal branches into PortalWorkspace
